+++ demos/default/next.config.js
@@ -78,6 +78,7 @@ module.exports = {
     includePaths: [path.join(__dirname, 'styles-sass-test')],
   },
   experimental: {
+    optimizeCss: true,
     images: {
       remotePatterns: [
         {